Updated the optimization algorithm. Significant decrease in tests.

+++ src/org/sag/gui/MutationPanel2.java
@@ -713,6 +713,64 @@ public class MutationPanel2 extends JPanel {
 		} else
 			return false;
 	}
+	
+	//check if 2 requests give the same decision for a mutant
+	private boolean checkResult(String req1, String req2, String mutant)
+	{
+		String request1 = "";
+		String request2 = "";
+		try
+		{
+			BufferedReader br = new BufferedReader(new FileReader(req1));
+			StringBuilder sb = new StringBuilder();
+			String line = br.readLine();
+			while(line != null)
+			{
+				sb.append(line);
+				sb.append(System.lineSeparator());
+				line = br.readLine();
+			}
+			request1 = sb.toString();
+			br.close();
+		}
+		catch(Exception e)
+		{
+			e.printStackTrace();
+			return false;
+		}
+		try
+		{
+			BufferedReader br = new BufferedReader(new FileReader(req2));
+			StringBuilder sb = new StringBuilder();
+			String line = br.readLine();
+			while(line != null)
+			{
+				sb.append(line);
+				sb.append(System.lineSeparator());
+				line = br.readLine();
+			}
+			request2 = sb.toString();
+			br.close();
+		}
+		catch(Exception e)
+		{
+			e.printStackTrace();
+			return false;
+		}
+		loadPolicy lp = new loadPolicy();
+		Policy m = lp.getPolicy(mutant);
+		if(m != null)
+		{
+			int res1 = lp.PolicyEvaluate(m, request1);
+			int res2 = lp.PolicyEvaluate(m, request2);
+			if(res1 == res2)
+				return true;
+			else
+				return false;
+		}
+		else
+			return false;
+	}
 
 	private void determineValidTests(
 			ArrayList<PolicySpreadSheetTestRecord> records,
@@ -761,12 +819,32 @@ public class MutationPanel2 extends JPanel {
 					valid.add(record);
 					validTests++;
 				} else {
-					System.err.println(req);
-					Runtime run = Runtime.getRuntime();
-					try {
-						run.exec("sudo rm " + req);
-					} catch (Exception e) {
-						System.err.println("File deletion failed");
+					if(type.contains("CRC"))
+					{
+						for(PolicyMutant m : mutator.getMutantList())
+						{
+							if(m.getNumber().contains("CRC"))
+							{
+								mutantPath = m.getMutantFilePath();
+								if(validation(policyPath, mutantPath, req))
+								{
+									valid.add(record);
+									validTests++;
+								}
+								else
+									continue;
+							}
+						}
+					}
+					if(!valid.contains(record))
+					{
+						System.err.println(req);
+						Runtime run = Runtime.getRuntime();
+						try {
+							run.exec("sudo rm " + req);
+						} catch (Exception e) {
+							System.err.println("File deletion failed");
+						}
 					}
 				}
 			}
@@ -1164,7 +1242,7 @@ public class MutationPanel2 extends JPanel {
 							&& mnum.compareTo(testNum) != 0) {
 						String rem = testPanel
 								.getTestOutputDestination("_MutationTests")
-								+ File.separator + mutant;
+								+ File.separator + m.getRequestFile();
 						Runtime run = Runtime.getRuntime();
 						try {
 							run.exec("sudo rm" + rem);
@@ -1205,7 +1283,7 @@ public class MutationPanel2 extends JPanel {
 							&& mnum.compareTo(testNum) != 0) {
 						String rem = testPanel
 								.getTestOutputDestination("_MutationTests")
-								+ File.separator + mutant;
+								+ File.separator + m.getRequestFile();
 						Runtime run = Runtime.getRuntime();
 						try {
 							run.exec("sudo rm" + rem);
@@ -1224,7 +1302,7 @@ public class MutationPanel2 extends JPanel {
 
 		i = 0;
 		j = i + 1;
-		while (j < test.length && i <= test.length)// semi-iterative bottom up
+		while (j < test.length && i < test.length)// semi-iterative bottom up
 		{
 			if (test[i] != null) {
 				PolicySpreadSheetTestRecord psstr = test[i];
@@ -1244,7 +1322,7 @@ public class MutationPanel2 extends JPanel {
 							&& mnum.compareTo(testNum) != 0) {
 						String rem = testPanel
 								.getTestOutputDestination("_MutationTests")
-								+ File.separator + mutant;
+								+ File.separator + m.getRequestFile();
 						Runtime run = Runtime.getRuntime();
 						try {
 							run.exec("sudo rm" + rem);
@@ -1286,7 +1364,7 @@ public class MutationPanel2 extends JPanel {
 							&& mnum.compareTo(testNum) != 0) {
 						String rem = testPanel
 								.getTestOutputDestination("_MutationTests")
-								+ File.separator + mutant;
+								+ File.separator + m.getRequestFile();
 						Runtime run = Runtime.getRuntime();
 						try {
 							run.exec("sudo rm" + rem);
@@ -1305,6 +1383,183 @@ public class MutationPanel2 extends JPanel {
 			} else
 				l--;
 		}
+		i = 0;
+		j = test.length - 1;
+		k = i;
+		l = j;
+		
+		while (i < test.length && j >= 0) {
+			if (i == j) {
+				i++;
+				j--;
+				continue;
+			}
+			if (test[i] != null) {
+				PolicySpreadSheetTestRecord psstr = test[i];
+				String req = testPanel
+						.getTestOutputDestination("_MutationTests")
+						+ File.separator + psstr.getRequestFile();
+				String testNum = psstr.getNumber().substring(
+						psstr.getNumber().indexOf(' ') + 1,
+						psstr.getNumber().length());
+				if (test[j] != null) {
+					PolicySpreadSheetTestRecord m = test[j];
+					String mnum = m.getNumber().substring(
+							m.getNumber().indexOf(' ') + 1,
+							m.getNumber().length());
+					String mutant = mutator.getMutantFileName(mnum);
+					String req2 = testPanel.getTestOutputDestination("_MutationTests")
+							+ File.separator + m.getRequestFile();
+					if (checkResult(req, req2, mutant) && mnum.compareTo(testNum) != 0) {
+						String rem = testPanel
+								.getTestOutputDestination("_MutationTests")
+								+ File.separator + m.getRequestFile();
+						Runtime run = Runtime.getRuntime();
+						try {
+							run.exec("sudo rm" + rem);
+						} catch (Exception e) {
+							e.printStackTrace();
+						}
+						test[j] = null;
+						j--;
+					} else
+						i++;
+				} else
+					j--;
+			} else
+				i++;
+		}
+		
+		while (l >= 0 && k < test.length) {
+			if (k == l) {
+				k++;
+				l--;
+				continue;
+			}
+			if (test[l] != null) {
+				PolicySpreadSheetTestRecord psstr = test[l];
+				String req = testPanel
+						.getTestOutputDestination("_MutationTests")
+						+ File.separator + psstr.getRequestFile();
+				String testNum = psstr.getNumber().substring(
+						psstr.getNumber().indexOf(' ') + 1,
+						psstr.getNumber().length());
+				if (test[k] != null) {
+					PolicySpreadSheetTestRecord m = test[k];
+					String mnum = m.getNumber().substring(
+							m.getNumber().indexOf(' ') + 1,
+							m.getNumber().length());
+					String req2 = testPanel.getTestOutputDestination("_MutationTests")
+							+ File.separator + m.getRequestFile();
+					String mutant = mutator.getMutantFileName(mnum);
+					if (checkResult(req, req2, mutant) && mnum.compareTo(testNum) != 0) {
+						String rem = testPanel
+								.getTestOutputDestination("_MutationTests")
+								+ File.separator + m.getRequestFile();
+						Runtime run = Runtime.getRuntime();
+						try {
+							run.exec("sudo rm" + rem);
+						} catch (Exception e) {
+							e.printStackTrace();
+						}
+						test[k] = null;
+						k++;
+					} else
+						l--;
+				} else
+					k++;
+			} else
+				l--;
+		}
+		
+		i = 0;
+		j = i + 1;
+		while (j < test.length && i < test.length)// semi-iterative bottom up
+		{
+			if (test[i] != null) {
+				PolicySpreadSheetTestRecord psstr = test[i];
+				String req = testPanel
+						.getTestOutputDestination("_MutationTests")
+						+ File.separator + psstr.getRequestFile();
+				String testNum = psstr.getNumber().substring(
+						psstr.getNumber().indexOf(' ') + 1,
+						psstr.getNumber().length());
+				if (test[j] != null) {
+					PolicySpreadSheetTestRecord m = test[j];
+					String mnum = m.getNumber().substring(
+							m.getNumber().indexOf(' ') + 1,
+							m.getNumber().length());
+					String req2 = testPanel.getTestOutputDestination("_MutationTests")
+							+ File.separator + m.getRequestFile();
+					String mutant = mutator.getMutantFileName(mnum);
+					if (checkResult(req, req2, mutant)
+							&& mnum.compareTo(testNum) != 0) {
+						String rem = testPanel
+								.getTestOutputDestination("_MutationTests")
+								+ File.separator + m.getRequestFile();
+						Runtime run = Runtime.getRuntime();
+						try {
+							run.exec("sudo rm" + rem);
+						} catch (Exception e) {
+							e.printStackTrace();
+						}
+						test[j] = null;
+						j++;
+					} else {
+						i = j;
+						j = i + 1;
+						continue;
+					}
+				} else
+					j++;
+			} else
+				i++;
+		}
+		
+		l = test.length - 1;
+		k = l - 1;
+		while (l >= 0 && k >= 0)// semi-iterative top down
+		{
+			if (test[l] != null) {
+				PolicySpreadSheetTestRecord psstr = test[l];
+				String req = testPanel
+						.getTestOutputDestination("_MutationTests")
+						+ File.separator + psstr.getRequestFile();
+				String testNum = psstr.getNumber().substring(
+						psstr.getNumber().indexOf(' ') + 1,
+						psstr.getNumber().length());
+				if (test[k] != null) {
+					PolicySpreadSheetTestRecord m = test[k];
+					String mnum = m.getNumber().substring(
+							m.getNumber().indexOf(' ') + 1,
+							m.getNumber().length());
+					String req2 = testPanel.getTestOutputDestination("_MutationTests")
+							+ File.separator + m.getRequestFile();
+					String mutant = mutator.getMutantFileName(mnum);
+					if (checkResult(req, req2, mutant)
+							&& mnum.compareTo(testNum) != 0) {
+						String rem = testPanel
+								.getTestOutputDestination("_MutationTests")
+								+ File.separator + m.getRequestFile();
+						Runtime run = Runtime.getRuntime();
+						try {
+							run.exec("sudo rm" + rem);
+						} catch (Exception e) {
+							e.printStackTrace();
+						}
+						test[k] = null;
+						k--;
+					} else {
+						l = k;
+						k = l - 1;
+						continue;
+					}
+				} else
+					k--;
+			} else
+				l--;
+		}
+		
 		valid = new ArrayList<PolicySpreadSheetTestRecord>();
 		for (PolicySpreadSheetTestRecord r : test)
 			if (r != null)
